pyproject.toml
[project]
name = "suzieq-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    "mcp[cli]>=1.6.0",
]
